[FAB2023-954] refactor: 사용자 동기화 API 변경
