+++ project/ovp/server/src/main/java/com/mobigen/ovp/common/CommonController.java
@@ -1,4 +1,4 @@
-package com.mobigen.ovp.user;
+package com.mobigen.ovp.common;
 
 import com.mobigen.framework.result.annotation.ResponseJsonResult;
 import lombok.RequiredArgsConstructor;
@@ -11,9 +11,9 @@ import org.springframework.web.bind.annotation.RestController;
 @Slf4j
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/user")
-public class UserController {
-    private final UserService userService;
+@RequestMapping("/api/common")
+public class CommonController {
+    private final CommonService commonService;
 
     /**
      * 사용자 동기화 > Open metadata의 User와 DB를 동기화 하는 API
@@ -23,6 +23,6 @@ public class UserController {
     @ResponseJsonResult(errorMessage = "동기화 실패")
     @PostMapping("/synchronization")
     public Object login() throws Exception {
-        return userService.synchronizeUser();
+        return commonService.synchronizeUser();
     }
 }
